requirements.txt
pymysql==1.1.1
rich==13.7.1
psycopg2-binary==2.9.9
paddlepaddle==2.6.2
paddleocr==2.7.3
paddlenlp==2.6.1
opencv-python==4.6.0.66
Pillow==10.3.0
numpy==1.24.3